+++ railties/lib/rails/application.rb
@@ -154,6 +154,14 @@ module Rails
       self
     end
 
+    # Load the application runner and invoke the registered hooks.
+    # Check <tt>Rails::Railtie.runner</tt> for more info.
+    def load_runner(app=self)
+      initialize_runner
+      super
+      self
+    end
+
     # Rails.application.env_config stores some of the Rails initial environment parameters.
     # Currently stores:
     #
@@ -303,6 +311,9 @@ module Rails
       require "pp"
       require "rails/console/app"
       require "rails/console/helpers"
+    end
+
+    def initialize_runner #:nodoc:
     end
 
     def build_original_fullpath(env)